+++ changelog.md
@@ -8,7 +8,7 @@ Codion Change Log
 ### is.codion.common.model
 - DefaultConditionModel.DefaultSetCondition refactored.
 - MultiSelection.Indexes now extends Value.
-- MultiSelection.Index removed, Value used instead.
+- MultiSelection.Items now extends Value.
 
 ## 0.18.26
 ### is.codion.common.model